+++ frontend/src/features/workflowStudio/inspector/AgentEditor.tsx
@@ -12,7 +12,7 @@ import { useUiStore } from '../../../stores/uiStore'
 import { AgentVersionsDialog } from './AgentVersionsDialog'
 import styles from './AgentsPanel.module.css'
 
-const runtimes: AgentRuntime[] = ['pi', 'openclaw', 'velites']
+const runtimes: AgentRuntime[] = ['pi', 'velites']
 const toolOptions = ['read', 'write', 'bash']
 
 type Props = {